billing_api/requirements.txt
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic[email]
fastapi==0.115.4
asyncpg==0.29.0
typer==0.12.5
click==8.1.7
uvicorn==0.15.0
aiohttp==3.10.10
backoff==1.11.1
ujson==5.10.0
json-logging==1.3.0
PyJWT==2.6.0
python-multipart==0.0.6
sqlalchemy==2.0.23
aioredis==2.0.1
pytz==2024.2